+++ src/Blameable/Mapping/Driver/Annotation.php
@@ -49,13 +49,16 @@ class Annotation extends AbstractAnnotationDriver
         $class = $this->getMetaReflectionClass($meta);
         // property annotations
         foreach ($class->getProperties() as $property) {
-            if ($meta->isMappedSuperclass && !$property->isPrivate()
+            if (
+                isset($meta->associationMappings[$property->name]['inherited'])
+                || ($meta->isMappedSuperclass && !$property->isPrivate())
                 || $meta->isInheritedField($property->name)
-                || isset($meta->associationMappings[$property->name]['inherited'])
             ) {
                 continue;
             }
             if ($blameable = $this->reader->getPropertyAnnotation($property, self::BLAMEABLE)) {
+                assert($blameable instanceof Blameable);
+
                 $field = $property->getName();
 
                 if (!$meta->hasField($field) && !$meta->hasAssociation($field)) {
@@ -87,6 +90,8 @@ class Annotation extends AbstractAnnotationDriver
                         'value' => $blameable->value,
                     ];
                 }
+                // add the setter method for the field
+                $this->setSetterMethod($field, $blameable->setterMethod, $config);
                 // properties are unique and mapper checks that, no risk here
                 $config[$blameable->on][] = $field;
             }