+++ CHANGELOG.md
@@ -67,6 +67,9 @@ a release.
 
 ## [3.12.0] - 2023-07-08
 ### Added
+- Blameable/IpTraceable/SoftDeletable/Timestampable: Added functionality to use setter method instead of setting property values directly (#2644)
+
+### Added
 - Tree: `setSibling()` and `getSibling()` methods in the `Node` interface through the BC `@method` annotation
 - Tree: Support array of fields and directions in the `$sortByField` and `$direction` parameters at `AbstractTreeRepository::recover()`
 - Loggable: Support for composite identifiers